added proper typecasting to arrays and repairing tests

+++ heat/core/tests/test_manipulations.py
@@ -821,11 +821,13 @@ class TestManipulations(BasicTest):
         data = ht.array(tensor, split=None, device=ht_device)
         result, result_indices = ht.sort(data, axis=0, descending=True)
         expected, exp_indices = torch.sort(tensor, dim=0, descending=True)
+        exp_indices = exp_indices.int()
         self.assertTrue(torch.equal(result._DNDarray__array, expected))
         self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_indices))
 
         result, result_indices = ht.sort(data, axis=1, descending=True)
         expected, exp_indices = torch.sort(tensor, dim=1, descending=True)
+        exp_indices = exp_indices.int()
         self.assertTrue(torch.equal(result._DNDarray__array, expected))
         self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_indices))
 
@@ -834,6 +836,7 @@ class TestManipulations(BasicTest):
         exp_axis_zero = torch.arange(size, device=device).reshape(1, size)
         exp_indices = torch.tensor([[rank] * size], device=device)
         result, result_indices = ht.sort(data, descending=True, axis=0)
+        exp_indices = exp_indices.int()
         self.assertTrue(torch.equal(result._DNDarray__array, exp_axis_zero))
         self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_indices))
 
@@ -842,6 +845,7 @@ class TestManipulations(BasicTest):
         )
         result, result_indices = ht.sort(data, descending=True, axis=1)
         self.assertTrue(torch.equal(result._DNDarray__array, exp_axis_one))
+        exp_indices = exp_indices.int()
         self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_indices))
 
         result1 = ht.sort(data, axis=1, descending=True)
@@ -857,12 +861,13 @@ class TestManipulations(BasicTest):
         self.assertTrue(torch.equal(result._DNDarray__array, exp_axis_zero))
         # comparison value is only true on CPU
         if result_indices._DNDarray__array.is_cuda is False:
+            indices_axis_zero = indices_axis_zero.int()
             self.assertTrue(torch.equal(result_indices._DNDarray__array, indices_axis_zero))
 
         exp_axis_one = torch.tensor(size - rank - 1, device=device).repeat(size).reshape(size, 1)
         result, result_indices = ht.sort(data, descending=True, axis=1)
         self.assertTrue(torch.equal(result._DNDarray__array, exp_axis_one))
-        self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_axis_one))
+        self.assertTrue(torch.equal(result_indices._DNDarray__array, exp_axis_one.int()))
 
         tensor = torch.tensor(
             [
@@ -1079,7 +1084,7 @@ class TestManipulations(BasicTest):
         self.assertEqual(inv.split, None)
         self.assertEqual(inv.dtype, data_split_none.dtype)
         self.assertEqual(inv.device, data_split_none.device)
-        self.assertTrue(torch.equal(inv._DNDarray__array, exp_inv))
+        self.assertTrue(torch.equal(inv._DNDarray__array, exp_inv.int()))
 
         data_split_zero = ht.array(torch_array, split=0, device=ht_device)
         res, inv = ht.unique(data_split_zero, return_inverse=True, sorted=True)
